Amended getCountOfStudies method to return Long.intValue()
git-svn-id: svn://localhost@4972 35ee9b83-dd2c-47f4-99a2-c706670c48ce

+++ ark-common/src/main/java/au/org/theark/core/dao/StudyDao.java
@@ -1037,7 +1037,7 @@ public class StudyDao<T> extends HibernateSessionDao implements IStudyDao {
 	
 	public int getCountOfStudies() {
 		int total = 0;
-		total = ((Integer) getSession().createQuery("select count(*) from Study").iterate().next()).intValue();
+		total = ((Long) getSession().createQuery("select count(*) from Study").iterate().next()).intValue();
 		return total;
 	}
 }
